split off a bunch of the runner/yaml/weight init code common to transformers
