+++ tt-train/sources/ttml/models/common/transformer_common.cpp
@@ -1,0 +1,91 @@
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "transformer_common.hpp"
+
+#include "yaml-cpp/yaml.h"
+
+namespace ttml::models::common::transformer {
+
+autograd::TensorPtr memory_efficient_runner(
+    auto&& forward_impl, const autograd::TensorPtr& input, const autograd::TensorPtr& mask) {
+    if (autograd::ctx().get_gradient_mode() == autograd::GradMode::DISABLED) {
+        return forward_impl(input, mask);
+    }
+
+    // make a copy of a generator before running forward pass
+    auto generator = autograd::ctx().get_generator();
+
+    // running forward pass
+    autograd::TensorPtr out;
+    {
+        auto scoped = ttml::core::Scoped(
+            []() { autograd::ctx().set_gradient_mode(autograd::GradMode::DISABLED); },
+            []() { autograd::ctx().set_gradient_mode(autograd::GradMode::ENABLED); });
+        out = forward_impl(input, mask);
+    }
+
+    // define grad function and copy generator (in the state before forward pass)
+    autograd::GradFunction grad = [input, mask, out, &forward_impl, generator]() {
+        // detach input from existing graph
+        auto input_detached = autograd::create_tensor(input->get_value());
+        // run forward pass again
+        autograd::TensorPtr output;
+        {
+            // set generator to the state before forward pass during construction
+            // restore generator state after grad function is executed
+            auto scoped = ttml::core::Scoped(
+                [&generator]() { autograd::ctx().set_generator(generator); },
+                [generator = autograd::ctx().get_generator()]() { autograd::ctx().set_generator(generator); });
+            output = forward_impl(input_detached, mask);
+        }
+        // use gradients from new output
+        output->set_grad(out->get_grad());
+        output->backward();
+        // reuse gradients from detached input
+        input->add_grad(input_detached->get_grad());
+    };
+
+    auto links = autograd::get_links(input);
+    out->set_node(autograd::ctx().add_backward_node(std::move(grad), links));
+    return out;
+}
+
+void weights_initialization(autograd::ModuleBase& model) {
+    auto params = model.parameters();
+    for (auto& [name, tensor_ptr] : params) {
+        const auto& tensor = tensor_ptr->get_value();
+        if (name.find("weight") != std::string::npos) {
+            init::normal_init(tensor_ptr, tensor.get_logical_shape(), {0.F, 0.02F});
+        } else if (name.find("bias") != std::string::npos) {
+            init::constant_init(tensor_ptr, tensor.get_logical_shape(), 0.F);
+        }
+    }
+}
+
+RunnerType read_runner_type(const YAML::Node& config) {
+    auto runner_type_str = config["runner_type"].as<std::string>("default");
+    if (runner_type_str == "default") {
+        return RunnerType::Default;
+    } else if (runner_type_str == "memory_efficient") {
+        return RunnerType::MemoryEfficient;
+    } else {
+        throw std::runtime_error(fmt::format(
+            "Unknown runner type: {}. Supported runner types [default, memory_efficient]", runner_type_str));
+    }
+}
+
+WeightTyingType read_weight_tying_type(const YAML::Node& config) {
+    auto weight_tying_str = config["weight_tying"].as<std::string>("disabled");
+    if (weight_tying_str == "disabled") {
+        return WeightTyingType::Disabled;
+    } else if (weight_tying_str == "enabled") {
+        return WeightTyingType::Enabled;
+    } else {
+        throw std::runtime_error(fmt::format(
+            "Unknown weight tying type: {}. Supported weight tying types [disabled, enabled]", weight_tying_str));
+    }
+}
+
+}  // namespace ttml::models::common::transformer
