+++ tt-train/sources/ttml/models/llama.cpp
@@ -4,78 +4,14 @@
 
 #include "llama.hpp"
 
-#include "autograd/graph_utils.hpp"
 #include "autograd/tensor.hpp"
-#include "core/scoped.hpp"
-#include "init/tensor_initializers.hpp"
 #include "modules/embedding_module.hpp"
 #include "modules/llama_block.hpp"
 #include "modules/rms_norm_module.hpp"
-#include "modules/rotary_embedding.hpp"
 #include "ops/rope_op.hpp"
 #include "ops/unary_ops.hpp"
 
 namespace ttml::models::llama {
-
-namespace {
-
-autograd::TensorPtr memory_efficient_runner(
-    auto&& forward_impl, const autograd::TensorPtr& input, const autograd::TensorPtr& mask) {
-    if (autograd::ctx().get_gradient_mode() == autograd::GradMode::DISABLED) {
-        return forward_impl(input, mask);
-    }
-
-    // make a copy of a generator before running forward pass
-    auto generator = autograd::ctx().get_generator();
-
-    // running forward pass
-    autograd::TensorPtr out;
-    {
-        auto scoped = ttml::core::Scoped(
-            []() { autograd::ctx().set_gradient_mode(autograd::GradMode::DISABLED); },
-            []() { autograd::ctx().set_gradient_mode(autograd::GradMode::ENABLED); });
-        out = forward_impl(input, mask);
-    }
-
-    // define grad function and copy generator (in the state before forward pass)
-    autograd::GradFunction grad = [input, mask, out, &forward_impl, generator]() {
-        // detach input from existing graph
-        auto input_detached = autograd::create_tensor(input->get_value());
-        // run forward pass again
-        autograd::TensorPtr output;
-        {
-            // set generator to the state before forward pass during construction
-            // restore generator state after grad function is executed
-            auto scoped = ttml::core::Scoped(
-                [&generator]() { autograd::ctx().set_generator(generator); },
-                [generator = autograd::ctx().get_generator()]() { autograd::ctx().set_generator(generator); });
-            output = forward_impl(input_detached, mask);
-        }
-        // use gradients from new output
-        output->set_grad(out->get_grad());
-        output->backward();
-        // reuse gradients from detached input
-        input->add_grad(input_detached->get_grad());
-    };
-
-    auto links = autograd::get_links(input);
-    out->set_node(autograd::ctx().add_backward_node(std::move(grad), links));
-    return out;
-}
-
-void weights_initialization(Llama& model) {
-    auto params = model.parameters();
-    for (auto& [name, tensor_ptr] : params) {
-        const auto& tensor = tensor_ptr->get_value();
-        if (name.find("weight") != std::string::npos) {
-            init::normal_init(tensor_ptr, tensor.get_logical_shape(), {0.F, 0.02F});
-        } else if (name.find("bias") != std::string::npos) {
-            init::constant_init(tensor_ptr, tensor.get_logical_shape(), 0.F);
-        }
-    }
-}
-
-}  // namespace
 
 Llama::Llama(const LlamaConfig& config) {
     uint32_t vocab_size = config.vocab_size;
@@ -136,7 +72,7 @@ Llama::Llama(const LlamaConfig& config) {
     register_module(ln_fc, "ln_fc");
     register_module(fc, "fc");
 
-    weights_initialization(*this);
+    common::transformer::weights_initialization(*this);
 }
 
 ttml::autograd::TensorPtr Llama::operator()(const ttml::autograd::TensorPtr& x, const ttml::autograd::TensorPtr& mask) {
@@ -144,7 +80,7 @@ ttml::autograd::TensorPtr Llama::operator()(const ttml::autograd::TensorPtr& x, 
     auto out = tok_emb_out;  // llama does positional embedding in the attention blocks
     for (auto& block : blocks) {
         if (runner_type == RunnerType::MemoryEfficient) {
-            out = memory_efficient_runner(*block, out, mask);
+            out = common::transformer::memory_efficient_runner(*block, out, mask);
         } else if (runner_type == RunnerType::Default) {
             out = (*block)(out, mask);
         } else {
@@ -157,30 +93,6 @@ ttml::autograd::TensorPtr Llama::operator()(const ttml::autograd::TensorPtr& x, 
     return log_softmax;
 }
 
-RunnerType read_runner_type(const YAML::Node& config) {
-    auto runner_type_str = config["runner_type"].as<std::string>("default");
-    if (runner_type_str == "default") {
-        return RunnerType::Default;
-    } else if (runner_type_str == "memory_efficient") {
-        return RunnerType::MemoryEfficient;
-    } else {
-        throw std::runtime_error(fmt::format(
-            "Unknown runner type: {}. Supported runner types [default, memory_efficient]", runner_type_str));
-    }
-}
-
-WeightTyingType read_weight_tying_type(const YAML::Node& config) {
-    auto weight_tying_str = config["weight_tying"].as<std::string>("disabled");
-    if (weight_tying_str == "disabled") {
-        return WeightTyingType::Disabled;
-    } else if (weight_tying_str == "enabled") {
-        return WeightTyingType::Enabled;
-    } else {
-        throw std::runtime_error(fmt::format(
-            "Unknown weight tying type: {}. Supported weight tying types [disabled, enabled]", weight_tying_str));
-    }
-}
-
 LlamaConfig read_config(const YAML::Node& config) {
     LlamaConfig llama_config;
     llama_config.num_heads = config["num_heads"].as<uint32_t>();
@@ -190,8 +102,8 @@ LlamaConfig read_config(const YAML::Node& config) {
     llama_config.num_blocks = config["num_blocks"].as<uint32_t>();
     llama_config.vocab_size = config["vocab_size"].as<uint32_t>();
     llama_config.max_sequence_length = config["max_sequence_length"].as<uint32_t>();
-    llama_config.runner_type = read_runner_type(config);
-    llama_config.weight_tying = read_weight_tying_type(config);
+    llama_config.runner_type = common::transformer::read_runner_type(config);
+    llama_config.weight_tying = common::transformer::read_weight_tying_type(config);
 
     return llama_config;
 }
