+++ tt-train/sources/ttml/models/llama.hpp
@@ -6,26 +6,20 @@
 
 #include "autograd/module_base.hpp"
 #include "autograd/tensor.hpp"
+#include "common/transformer_common.hpp"
 #include "modules/llama_block.hpp"
 #include "ops/rope_op.hpp"
 #include "yaml-cpp/yaml.h"
 
 namespace ttml::models::llama {
 
-enum class RunnerType {
-    MemoryEfficient,
-    Default,
-};
-
-enum class WeightTyingType {
-    Disabled,
-    Enabled,
-};
+using RunnerType = common::transformer::RunnerType;
+using WeightTyingType = common::transformer::WeightTyingType;
 
 struct LlamaConfig {
     uint32_t num_heads = 8;
     uint32_t num_groups = 2;
-    uint32_t embedding_dim = 384;
+    uint32_t embedding_dim = 384;  // embedding dimension, must be divisible by num_heads
     float dropout_prob = 0.0F;
     uint32_t num_blocks = 6;
     uint32_t vocab_size = 256;
